--- src/components/removeBunny.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export default function RemoveBunny({ onRemove },{ bunny }) {
-    return (
-        <button onClick={() => onRemove(bunny)}>X</button>
-    );
-}
